adding updates to the merge overlap pipeline, allowing for user-based plotting permissions
