+++ src/nextflow/nextflow.config
@@ -1,5 +1,6 @@
 params {
 	help = false
+	verbose = false
 	read = "single"
 	fasta = ""
 	workdir = ""
@@ -11,5 +12,5 @@ params {
 	toolpath = "src/modules/"
 	readlength = 100
 	kmer_length = 20
-	clustering_threshold = 0.5
+	clustering_threshold = 0.2
 }
